requirements.txt
retinaface
torch
opencv-python
pillow
tqdm